requirements.txt
fastapi
uvicorn
yfinance
pandas
pandas-ta-classic
numpy
scikit-learn
tensorflow
transformers
torch
plotly
streamlit
google-generativeai